Moved BO GPflux integration test to parent folder test, formatting
