+++ trieste/acquisition/sampler.py
@@ -354,10 +354,8 @@ class DeepGaussianProcessSampler(Sampler):
         # Each element of _eps_list is essentially a lazy constant. It is declared and assigned an
         # empty tensor here, and populated on the first call to sample
         self._eps_list = [
-            tf.Variable(
-                tf.ones([sample_size, 0], dtype=tf.float64), shape=[sample_size, None]
-            )
-        ]*len(model.model_gpflux.f_layers)
+            tf.Variable(tf.ones([sample_size, 0], dtype=tf.float64), shape=[sample_size, None])
+        ] * len(model.model_gpflux.f_layers)
 
     def sample(self, at: TensorType) -> TensorType:
         """
@@ -379,12 +377,10 @@ class DeepGaussianProcessSampler(Sampler):
 
             if not eps_is_populated:
                 self._eps_list[i].assign(
-                    tf.random.normal(
-                        [self._sample_size, tf.shape(mean)[-1]], dtype=tf.float64
-                    )
+                    tf.random.normal([self._sample_size, tf.shape(mean)[-1]], dtype=tf.float64)
                 )
 
-            samples = mean + tf.sqrt(var)*tf.cast(self._eps_list[i][:, None, :], var.dtype)
+            samples = mean + tf.sqrt(var) * tf.cast(self._eps_list[i][:, None, :], var.dtype)
 
         return samples
