+++ trieste/objectives/single_objectives.py
@@ -22,7 +22,6 @@ from __future__ import annotations
 
 import math
 
-import gpflow
 import tensorflow as tf
 
 from ..space import Box
